correctly use context managers to close the file only when the RegionFile class is configured to open files on demand.
as it turns out, experimenting with this option vs earlier code showed no change in the number of open filehandles. python will close the file when it garbage-collects the file object so we don't ever have to close the file.  on windows, opening the file for writing doesn't give an exclusive lock so we don't determine if minecraft is holding onto the file