Added distutils setup script written by user:sixohsix.
changed setup arguments, adding an explicit packages list and a package_dir override that allows me to keep the current folder structure. tests.py and run_regression_test.py are included in the dist as a result, and egg-info/SOURCES.txt is slightly mangled but the docs say it's not used for anything anyway.

plus, `setup.py develop` still works.