blocks are now defined in .yaml files. added a yaml library.
the .yaml format is an extended version of the one used by Minecraft X-Ray. fields are added for data variants, extra search terms, and lighting info. tables for converting blocks are now inferred from the block info: for each source block's name, it will search the destination's block info for names or aliases (aka) that match or contain the name.

the MCMaterials objects gained the [] operator, which can accept a block name, block id, or (id, data) pair.