Anvil format support (as it appears in 12w07)
A height of 256 is hard-coded for Anvil worlds (version == 19133). The AnvilChunk class emulates an old extended height chunk by swapping axes and stacking the chunk sections vertically.

The chunk compress/decompress thing really got in the way here. Compressed (unsaved) chunks should be stored separately instead of as a state of InfdevChunk.