Don't turn a MemoryError into a ChunkMalformed exception
In a number of places, all exceptions were caught and a ChunkMalformed
exception raised in their place.  Some code, such as dumpSigns,
continues after a ChunkMalformed exception.  In the case where the
ChunkMalformed exception was a result of a MemoryError, this will most
likely result in a flood of error messages because the MemoryError is
not likely to go away by itself.

This fix causes MemoryError exceptions to not be transformed into
ChunkMalformed exceptions, and hopefully covers all cases where this
was done.  Now, when memory is exhausted, the only detail that is
output is "Exception during command: MemoryError()", rather than the
previous, more detailed output, such as:

"""
ERROR:Malformed NBT data in file: r.4.0.mca index 3776 sector 552 length 1 format deflate ()
ERROR:Incorrect chunk format in file: r.4.0.mca index 3776 sector 552 length 1 format deflate (((MemoryError(),), <traceback object at 0x1C927B98>))
"""

but the particular data that was being processed at the time of the
MemoryError is not likely to be useful information anyway.